Double refers to f64 not f32 in c++, refactored logic to handle the requests correctly. Updated test constants to reflect new accuracy levels. Back to section 5 implementation

+++ src/main.rs
@@ -7,19 +7,19 @@ use std::io::prelude::*;
 
 fn main() {
     // Image Data
-    const ASPECT_RATIO: f32 = 16.0 / 9.0;
+    const ASPECT_RATIO: f64 = 16.0 / 9.0;
     let ray_color: RayColor = RayColor::new_rgb(0.0, 0.0, 0.0);
     
     let image_width: u32 = 1600;
-    let mut image_height: u32 = (image_width as f32 / ASPECT_RATIO) as u32;
+    let mut image_height: u32 = (image_width as f64 / ASPECT_RATIO) as u32;
     if image_height < 1 {
         image_height = 1;
     }
 
     // Camera Viewport Data
-    let focal_length: f32 = 1.0;
+    let focal_length: f64 = 1.0;
     let viewport_height = 2.0;
-    let viewport_width = viewport_height * (image_width as f32 / image_height as f32);
+    let viewport_width = viewport_height * (image_width as f64 / image_height as f64);
     let camera_center = Point3::new(0.0, 0.0, 0.0);
 
     // Calculate the vectors across the horizontal and down the vertical viewport edges.
@@ -27,8 +27,8 @@ fn main() {
     let viewport_v = RtVec3::new(0.0, -viewport_height, 0.0);
 
     // Calculate the horizontal and vertical delta vectors from pixel to pixel.
-    let pixel_delta_u = viewport_u / image_width as f32;
-    let pixel_delta_v = viewport_v / image_height as f32;
+    let pixel_delta_u = viewport_u / image_width as f64;
+    let pixel_delta_v = viewport_v / image_height as f64;
 
     // Calculate the location of the upper left pixel.
     let viewport_upper_left = camera_center - RtVec3::new(0.0, 0.0, focal_length) - viewport_u / 2.0 - viewport_v / 2.0;
@@ -51,6 +51,29 @@ fn main() {
     println!("Point at t=2.0: {:?}", point);
 }
 
+// fn hit_sphere(
+//     center: Point3,
+//     radius: f64,
+//     ray: Ray,
+// ) -> bool {
+//     // vec3 oc = center - r.origin();
+//     let oc = center - ray.origin();
+
+//     // auto a = dot(r.direction(), r.direction());
+//     let a = ray.direction().dot(ray.direction())
+
+//     // auto b = -2.0 * dot(r.direction(), oc);
+//     let b = -2.0 * ray.direction().dot(oc)
+
+//     // auto c = dot(oc, oc) - radius*radius;
+//     let c = oc.dot(oc) - radius * radius;
+
+//     // auto discriminant = b*b - 4*a*c;
+//     let discriminant = b * b - 4 * a * c;
+
+//     return (discriminant >= 0);
+// }
+
 fn build_file(
     image_width: u32,
     image_height: u32,
@@ -70,8 +93,8 @@ fn build_file(
     for pixel_h in 0..image_height {
         println!("Scanline's remaining: {:?} ", (image_height - pixel_h));
         for pixel_w in 0..image_width {
-            let pixel_center = pixel_00_loc + pixel_w as f32 * pixel_delta_u + pixel_h as f32 * pixel_delta_v;
-            // let pixel_center = pixel_00_loc + (pixel_h as f32 * image_height as f32) + (pixel_w as f32* image_width as f32);
+            let pixel_center = pixel_00_loc + pixel_w as f64 * pixel_delta_u + pixel_h as f64 * pixel_delta_v;
+            // let pixel_center = pixel_00_loc + (pixel_h as f64 * image_height as f64) + (pixel_w as f64* image_width as f64);
             let ray_direction = pixel_center - camera_center;
             let ray = Ray::new(camera_center, ray_direction);
